# Performance Optimization of Sparse Matrix Access and RNG Generation